+++ package.json
@@ -13,7 +13,8 @@
     "qs": "~6.5.1",
     "request": "~2.83.0",
     "rxjs": "~5.5.2",
-    "sift": "~5.0.0"
+    "sift": "~5.0.0",
+    "url-join": "~2.0.2"
   },
   "devDependencies": {
     "babel-preset-env": "~1.6.1",
