--- packages/kinvey-identity/package.json
@@ -1,24 +1,0 @@
-{
-  "private": true,
-  "version": "3.9.3-beta.7",
-  "name": "kinvey-identity",
-  "main": "src/index.js",
-  "dependencies": {
-    "es6-promise": "4.1.1",
-    "kinvey-client": "^3.9.3-beta.7",
-    "kinvey-errors": "^3.9.3-beta.7",
-    "lodash": "4.17.4",
-    "url-join": "2.0.2"
-  },
-  "devDependencies": {
-    "expect": "1.20.2",
-    "kinvey": "^3.9.3-beta.7",
-    "kinvey-request": "^3.9.3-beta.7",
-    "kinvey-user": "^3.9.3-beta.7",
-    "mocha": "4.0.1",
-    "nock": "9.0.28"
-  },
-  "scripts": {
-    "test": "mocha -s 250 test/**/*.spec.js"
-  }
-}
