--- packages/kinvey-identity/src/index.js
@@ -1,8 +1,0 @@
-const { SocialIdentity } = require('./enums');
-const { AuthorizationGrant, MobileIdentityConnect } = require('./mic');
-
-module.exports = {
-  SocialIdentity: SocialIdentity,
-  AuthorizationGrant: AuthorizationGrant,
-  MobileIdentityConnect: MobileIdentityConnect
-};
