+++ src/identity/enums.js
@@ -2,9 +2,8 @@
  * @private
  * Enum for Social Identities
  */
-const SocialIdentity = {
+export const SocialIdentity = {
   Kinvey: 'kinvey',
   MobileIdentityConnect: 'kinveyAuth',
 };
 Object.freeze(SocialIdentity);
-exports.SocialIdentity = SocialIdentity;
