+++ src/identity/index.js
@@ -1,0 +1,2 @@
+export * from './enums';
+export * from './mic';
